Backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
requests
passlib[bcrypt]